+++ src/Roslyn.Suppressions/Roslyn.Suppressions.UnitTests/RoslynSettingsFileInfoTests.cs
@@ -22,6 +22,7 @@ using System.Globalization;
 using System.IO;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using SonarLint.VisualStudio.Integration.UnitTests.Helpers;
 using SonarLint.VisualStudio.Roslyn.Suppressions.SettingsFile;
 
 namespace SonarLint.VisualStudio.Roslyn.Suppressions.UnitTests
@@ -41,12 +42,12 @@ namespace SonarLint.VisualStudio.Roslyn.Suppressions.UnitTests
 
             //This is to make sure normalising the keys done correctly with invariant culture
             //https://en.wikipedia.org/wiki/Dotted_and_dotless_I 
-            using (new TemporaryCultureSwitch(new CultureInfo("tr-TR")))
-            {
-                var actualPath = RoslynSettingsFileInfo.GetSettingsFilePath(projectKey);
+            using var scope = new TurkishCultureScope();
 
-                actualPath.Should().Be(expectedPath);
-            }
+            var actualPath = RoslynSettingsFileInfo.GetSettingsFilePath(projectKey);
+
+            actualPath.Should().Be(expectedPath);
+            
         }
     }
 }
